[misc] refactor all object type checks in single file
All the files in src/lib/utils that handled type checking have been merged into one file :
src/lib/utils/check-object-type.js

This file now includes following functions :
isArray(),
isDate(),
isFunction(),
isNumber(),
isObject(),
isString(),
isUndefined()
