+++ src/lib/locale/calendar.js
@@ -7,7 +7,7 @@ export var defaultCalendar = {
     sameElse: 'L',
 };
 
-import isFunction from '../utils/is-function';
+import { isFunction } from '../utils/check-object-type';
 
 export function calendar(key, mom, now) {
     var output = this._calendar[key] || this._calendar['sameElse'];
