+++ src/lib/locale/relative.js
@@ -17,7 +17,7 @@ export var defaultRelativeTime = {
     yy: '%d years',
 };
 
-import isFunction from '../utils/is-function';
+import { isFunction } from '../utils/check-object-type';
 
 export function relativeTime(number, withoutSuffix, string, isFuture) {
     var output = this._relativeTime[string];
