+++ src/lib/moment/format.js
@@ -1,6 +1,6 @@
 import { formatMoment } from '../format/format';
 import { hooks } from '../utils/hooks';
-import isFunction from '../utils/is-function';
+import { isFunction } from '../utils/check-object-type';
 
 hooks.defaultFormat = 'YYYY-MM-DDTHH:mm:ssZ';
 hooks.defaultFormatUtc = 'YYYY-MM-DDTHH:mm:ss[Z]';
