+++ src/lib/moment/get-set.js
@@ -1,7 +1,7 @@
 import { normalizeUnits, normalizeObjectUnits } from '../units/aliases';
 import { getPrioritizedUnits } from '../units/priorities';
 import { hooks } from '../utils/hooks';
-import isFunction from '../utils/is-function';
+import { isFunction } from '../utils/check-object-type';
 import { isLeapYear } from '../units/year';
 
 export function makeGetSet(unit, keepTime) {
