+++ src/lib/units/month.js
@@ -13,8 +13,7 @@ import { addParseToken } from '../parse/token';
 import { hooks } from '../utils/hooks';
 import { MONTH } from './constants';
 import toInt from '../utils/to-int';
-import isArray from '../utils/is-array';
-import isNumber from '../utils/is-number';
+import { isArray, isNumber } from '../utils/check-object-type';
 import mod from '../utils/mod';
 import indexOf from '../utils/index-of';
 import { createUTC } from '../create/utc';
