--- src/lib/utils/is-array.js
@@ -1,6 +1,0 @@
-export default function isArray(input) {
-    return (
-        input instanceof Array ||
-        Object.prototype.toString.call(input) === '[object Array]'
-    );
-}
