--- src/lib/utils/is-date.js
@@ -1,6 +1,0 @@
-export default function isDate(input) {
-    return (
-        input instanceof Date ||
-        Object.prototype.toString.call(input) === '[object Date]'
-    );
-}
