--- src/lib/utils/is-function.js
@@ -1,6 +1,0 @@
-export default function isFunction(input) {
-    return (
-        (typeof Function !== 'undefined' && input instanceof Function) ||
-        Object.prototype.toString.call(input) === '[object Function]'
-    );
-}
