--- src/lib/utils/is-number.js
@@ -1,6 +1,0 @@
-export default function isNumber(input) {
-    return (
-        typeof input === 'number' ||
-        Object.prototype.toString.call(input) === '[object Number]'
-    );
-}
