--- src/lib/utils/is-string.js
@@ -1,3 +1,0 @@
-export default function isString(input) {
-    return typeof input === 'string' || input instanceof String;
-}
