--- src/lib/utils/is-undefined.js
@@ -1,3 +1,0 @@
-export default function isUndefined(input) {
-    return input === void 0;
-}
